pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["hypelilar", "Kamyar Ghasemlou <hypebot@kamy.me>"]

[tool.poetry.dependencies]
python = "^3.8"
cv2utils = {git = "https://github.com/yigitbey/cv2utils.git"}
slack-bolt = "^1.4.1"
requests = "^2.25.1"
Pillow = "^8.1.0"
google-api-python-client = "^1.12.8"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"
Flask = "^1.1.2"
gunicorn = "^20.0.4"
slack-sdk = "^3.4.2"
toml = "^0.10.2"
pre-commit = "^2.11.1"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
